requirements.txt
PyYAML>=6.0
questionary>=2.0.0 